pyproject.toml
[tool.poetry]
name = "pytest_generator_new"
version = "0.1.0"
description = "Homework 5 grading, CAP-5771, Spring 2024"
authors = ["erlebach <gordon.erlebach@gmail.com>"]
readme = "README.md"
packages = [{include = "pytest_generator"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.3"
matplotlib = "^3.8.2"
pytest = "^8.0.0"
gradescope-utils = "^0.5.0"
pyyaml = "^6.0.1"
pytest-mock = "^3.12.0"
tqdm = "^4.66.2"


[tool.poetry.group.dev.dependencies]
pytest-json-report = "^1.5.0"
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#[tool.poetry.packages]
#packages = [ 
#{ include = "pytest_generator", from = "src" } 
#]

[tool.poetry.scripts]
test_utils = "pytest_generator.test_utils:main"
yaml_expand = "pytest_generator.yaml_expand:main"
test_generator = "pytest_generator.test_generator:main"
code_generator = "pytest_generator.template_code_generator:main"
